# api/openapi.yaml
openapi: 3.1.0
info:
  title: Latasya ERP API
  version: 1.0.0
  description: |
    JSON API for Latasya ERP — a bookkeeping web app for Indonesian transport businesses.
    
    ## Authentication
    
    Two authentication methods are supported:
    
    - **Cookie + CSRF**: For browser clients. Session cookie `session_id` + `X-CSRF-Token` header.
    - **Bearer Token**: For bots, MCP servers, Telegram integrations. `Authorization: Bearer lat_<token>`.
      Bearer tokens are scoped (subset of user capabilities) and revocable.
      Bearer requests skip CSRF validation.
    
    ## Money & Quantities
    
    - **Currency (IDR)**: All monetary amounts are strings representing integer IDR values.
      Example: `"50000"` = Rp 50.000. No decimals, no subunits.
    - **Quantities**: Decimal strings with up to 2 decimal places.
      Example: `"1.50"` = 1.5 units (stored internally as integer × 100).
    
    ## Pagination
    
    List endpoints return: `{"data": [...], "meta": {"page": 1, "per_page": 50, "total": 100, "total_pages": 2}}`
    
    Query params: `?page=1&per_page=50` (default 50, max 200).
    
    ## Errors
    
    All errors return: `{"error": "message", "code": "snake_case_code", "request_id": "...", "fields": {...}}`
    
    Common codes: `unauthorized`, `forbidden`, `not_found`, `validation_failed`, `conflict`, `rate_limited`, `invalid_token`.
servers:
  - url: /api/v1
    description: Latasya ERP API v1
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token created via /settings/api-tokens or POST /api/v1/api-tokens.
        Format: `lat_<32-base62-chars>`. Scoped to a subset of user capabilities.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie set on login. Requires X-CSRF-Token header for state-changing requests.
...
  schemas:
...
    Instant:
      type: string
      format: date-time
      description: UTC timestamp in ISO-8601 format.
      example: "2026-05-10T03:00:00Z"
...
    ErrorEnvelope:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "not authenticated"
        code:
          type: string
          description: Machine-readable error code.
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - invalid_token
            - password_change_required
            - idempotency_conflict
            - internal_error
          example: "unauthorized"
        request_id:
          type: string
          description: Request ID for tracing (from audit middleware).
          example: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4"
        fields:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors (only present for validation_failed).
          example:
            contact_id: "required"
            amount: "must be positive"
...
    APIToken:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        prefix:
          type: string
          description: First 8 chars of token (safe to display)
        scopes:
          type: array
          items:
            type: string
        expires_at:
          allOf:
            - $ref: '#/components/schemas/Instant'
          nullable: true
        last_used_at:
          allOf:
            - $ref: '#/components/schemas/Instant'
          nullable: true
        revoked_at:
          allOf:
            - $ref: '#/components/schemas/Instant'
          nullable: true
        created_at:
          $ref: '#/components/schemas/Instant'
    APITokenCreated:
      allOf:
        - $ref: '#/components/schemas/APIToken'
        - type: object
          properties:
            plaintext:
              type: string
              description: Full token value. Shown ONCE only. Store securely.
    APITokenInput:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
          description: Must be a subset of the requesting user's capabilities.
        expires_at:
          allOf:
            - $ref: '#/components/schemas/Instant'
          nullable: true
          description: Optional expiry. Null means no expiry.
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
paths:
...
  /api-tokens:
    get:
      operationId: listAPITokens
      summary: List current user's API tokens
      responses:
        '200':
          description: List of API tokens (never includes plaintext)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createAPIToken
      summary: Create a new API token (cookie auth only)
      description: |
        Creates a scoped Bearer token for bot/MCP/Telegram integrations.
        The plaintext token is returned ONCE in this response only.
        Bearer-authenticated requests are rejected (tokens cannot create tokens).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APITokenInput'
      responses:
        '201':
          description: Token created (plaintext shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/APITokenCreated'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api-tokens/{id}:
    delete:
      operationId: revokeAPIToken
      summary: Revoke an API token (cookie auth only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Token revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'